• Lots of renaming and cleaning and converting to and from formats... :(  => Bloody pixel formats!!  => Not even finished with it! >8'(

+++ BaseLib/ASMHelpers.h
@@ -10,7 +10,7 @@ namespace BaseLib {
 	// Win64 architecture
 	#include <memory.h>	// include regular memset/memcpy/etc.
 
-	#define log2f( a )			(3.3219280948873623478703194294894f * log( a ))
+	#define log2f( a )			(3.3219280948873623478703194294894f * float(log( a )))
 
 #else // !defined(_WIN64)
